Create "DockerWatcher" so we can track docker instances booted by our subinstances.
Basically, we stick ourselves between the /var/run/docker.sock and the one we
expose to the child docker. This lets us watch the commands it makes, so we can
shut down any containers it boots.

In theory, we could use this mechanism to enforce a true "docker-in-docker" security
model by insisting that children only expose subvolumes of the ones we exposed
to the parent docker, and that all other privileges are tightened only.